+++ app/Traits/HasMeta.php
@@ -41,10 +41,6 @@ trait HasMeta
         return $this;
     }
 
-    // public function hasMeta(string $key): bool
-    // {
-    //     return array_key_exists($key, $this->meta ?? []);
-    // }
     public function hasMeta(string $key): bool
     {
         return data_get($this->meta ?? [], $key) !== null;